+++ src/main/java/svenhjol/charm/module/coral_squids/CoralSquidEntity.java
@@ -107,9 +107,9 @@ public class CoralSquidEntity extends WaterAnimal {
             && pos.getY() < level.getSeaLevel()
             && coralBelow;
 
-        if (canSpawn) {
-            LogHelper.debug(CoralSquidEntity.class, "Can spawn coral squid at pos: " + pos);
-        }
+//        if (canSpawn) {
+//            LogHelper.debug(CoralSquidEntity.class, "Can spawn coral squid at pos: " + pos);
+//        }
 
         return canSpawn;
     }